+++ tests/amazon/aws/operators/test_sagemaker.py
@@ -4,6 +4,7 @@ import pytest
 from airflow.exceptions import AirflowException, TaskDeferred
 from airflow.providers.amazon.aws.hooks.sagemaker import SageMakerHook
 from airflow.providers.amazon.aws.operators import sagemaker
+from airflow.utils.timezone import datetime
 
 from astronomer.providers.amazon.aws.operators.sagemaker import (
     SageMakerProcessingOperatorAsync,
@@ -135,14 +136,89 @@ class TestSagemakerProcessingOperatorAsync:
     CHECK_INTERVAL = 5
     MAX_INGESTION_TIME = 60 * 60 * 24 * 7
 
+    @mock.patch("astronomer.providers.amazon.aws.operators.sagemaker.SageMakerProcessingOperatorAsync.defer")
+    @mock.patch.object(
+        SageMakerHook, "describe_processing_job", return_value={"ProcessingJobStatus": "Completed"}
+    )
     @mock.patch.object(
         SageMakerHook,
         "create_processing_job",
         return_value={"ProcessingJobArn": "test_arn", "ResponseMetadata": {"HTTPStatusCode": 200}},
     )
     @mock.patch.object(SageMakerHook, "count_processing_jobs_by_name", return_value=False)
-    def test_sagemakerprocessing_op_async(self, mock_hook, mock_processing, context):
-        """Assert SageMakerProcessingOperatorAsync deferred properlu"""
+    def test_sagemakerprocessing_op_async_complete_before_defer(
+        self, mock_hook, mock_processing, mock_describe, mock_defer, context
+    ):
+        task = SageMakerProcessingOperatorAsync(
+            config=CREATE_PROCESSING_PARAMS,
+            task_id=self.TASK_ID,
+            check_interval=self.CHECK_INTERVAL,
+            max_ingestion_time=self.MAX_INGESTION_TIME,
+        )
+        task.execute(context)
+
+        assert not mock_defer.called
+
+    @mock.patch("astronomer.providers.amazon.aws.operators.sagemaker.SageMakerProcessingOperatorAsync.defer")
+    @mock.patch.object(SageMakerHook, "describe_processing_job", return_value=Exception)
+    @mock.patch.object(
+        SageMakerHook,
+        "create_processing_job",
+        return_value={"ProcessingJobArn": "test_arn", "ResponseMetadata": {"HTTPStatusCode": 200}},
+    )
+    @mock.patch.object(SageMakerHook, "count_processing_jobs_by_name", return_value=False)
+    def test_sagemakerprocessing_op_async_encounter_exception_before_defer(
+        self, mock_hook, mock_processing, mock_describe, mock_defer, context
+    ):
+        task = SageMakerProcessingOperatorAsync(
+            config=CREATE_PROCESSING_PARAMS,
+            task_id=self.TASK_ID,
+            check_interval=self.CHECK_INTERVAL,
+            max_ingestion_time=self.MAX_INGESTION_TIME,
+        )
+
+        with pytest.raises(Exception):
+            task.execute(context)
+
+        assert not mock_defer.called
+
+    @mock.patch("astronomer.providers.amazon.aws.operators.sagemaker.SageMakerProcessingOperatorAsync.defer")
+    @mock.patch.object(
+        SageMakerHook,
+        "describe_processing_job",
+        return_value={"ProcessingJobStatus": "Failed", "FailureReason": "it just failed"},
+    )
+    @mock.patch.object(
+        SageMakerHook,
+        "create_processing_job",
+        return_value={"ProcessingJobArn": "test_arn", "ResponseMetadata": {"HTTPStatusCode": 200}},
+    )
+    @mock.patch.object(SageMakerHook, "count_processing_jobs_by_name", return_value=False)
+    def test_sagemakerprocessing_op_async_failed_before_defer(
+        self, mock_hook, mock_processing, mock_describe, mock_defer, context
+    ):
+        task = SageMakerProcessingOperatorAsync(
+            config=CREATE_PROCESSING_PARAMS,
+            task_id=self.TASK_ID,
+            check_interval=self.CHECK_INTERVAL,
+            max_ingestion_time=self.MAX_INGESTION_TIME,
+        )
+        with pytest.raises(AirflowException):
+            task.execute(context)
+
+        assert not mock_defer.called
+
+    @mock.patch.object(
+        SageMakerHook, "describe_processing_job", return_value={"ProcessingJobStatus": "InProgress"}
+    )
+    @mock.patch.object(
+        SageMakerHook,
+        "create_processing_job",
+        return_value={"ProcessingJobArn": "test_arn", "ResponseMetadata": {"HTTPStatusCode": 200}},
+    )
+    @mock.patch.object(SageMakerHook, "count_processing_jobs_by_name", return_value=False)
+    def test_sagemakerprocessing_op_async(self, mock_hook, mock_processing, mock_describe, context):
+        """Assert SageMakerProcessingOperatorAsync deferred properly"""
         task = SageMakerProcessingOperatorAsync(
             config=CREATE_PROCESSING_PARAMS,
             task_id=self.TASK_ID,
@@ -219,6 +295,13 @@ class TestSagemakerTransformOperatorAsync:
     CHECK_INTERVAL = 5
     MAX_INGESTION_TIME = 60 * 60 * 24 * 7
 
+    @mock.patch("astronomer.providers.amazon.aws.operators.sagemaker.SageMakerTransformOperatorAsync.defer")
+    @mock.patch.object(SageMakerHook, "describe_model", return_value={})
+    @mock.patch.object(
+        SageMakerHook,
+        "describe_transform_job",
+        return_value={"TransformJobStatus": "Failed", "FailureReason": "it just failed"},
+    )
     @mock.patch.object(
         SageMakerHook,
         "create_transform_job",
@@ -226,7 +309,77 @@ class TestSagemakerTransformOperatorAsync:
     )
     @mock.patch.object(SageMakerHook, "list_transform_jobs", return_value=[])
     @mock.patch.object(SageMakerHook, "create_model", return_value=None)
-    def test_sagemaker_transform_op_async(self, mock_hook, mock_transform_job, context):
+    def test_sagemaker_transform_op_async_failed_before_defer(
+        self,
+        mock_create_model,
+        mock_list_transform_jobs,
+        mock_transform_job,
+        mock_describe_transform_job,
+        mock_describe_model,
+        mock_defer,
+        context,
+    ):
+        task = SageMakerTransformOperatorAsync(
+            config=CONFIG,
+            task_id=self.TASK_ID,
+            check_if_job_exists=False,
+            check_interval=self.CHECK_INTERVAL,
+            max_ingestion_time=self.MAX_INGESTION_TIME,
+        )
+        with pytest.raises(AirflowException):
+            task.execute(context)
+        assert not mock_defer.called
+
+    @mock.patch("astronomer.providers.amazon.aws.operators.sagemaker.SageMakerTransformOperatorAsync.defer")
+    @mock.patch.object(SageMakerHook, "describe_model", return_value={})
+    @mock.patch.object(
+        SageMakerHook, "describe_transform_job", return_value={"TransformJobStatus": "Completed"}
+    )
+    @mock.patch.object(
+        SageMakerHook,
+        "create_transform_job",
+        return_value={"TransformJobArn": "test_arn", "ResponseMetadata": {"HTTPStatusCode": 200}},
+    )
+    @mock.patch.object(SageMakerHook, "list_transform_jobs", return_value=[])
+    @mock.patch.object(SageMakerHook, "create_model", return_value=None)
+    def test_sagemaker_transform_op_async_complete_before_defer(
+        self,
+        mock_create_model,
+        mock_list_transform_jobs,
+        mock_transform_job,
+        mock_describe_transform_job,
+        mock_describe_model,
+        mock_defer,
+        context,
+    ):
+        task = SageMakerTransformOperatorAsync(
+            config=CONFIG,
+            task_id=self.TASK_ID,
+            check_if_job_exists=False,
+            check_interval=self.CHECK_INTERVAL,
+            max_ingestion_time=self.MAX_INGESTION_TIME,
+        )
+        task.execute(context)
+        assert not mock_defer.called
+
+    @mock.patch.object(
+        SageMakerHook, "describe_transform_job", return_value={"TransformJobStatus": "InProgress"}
+    )
+    @mock.patch.object(
+        SageMakerHook,
+        "create_transform_job",
+        return_value={"TransformJobArn": "test_arn", "ResponseMetadata": {"HTTPStatusCode": 200}},
+    )
+    @mock.patch.object(SageMakerHook, "list_transform_jobs", return_value=[])
+    @mock.patch.object(SageMakerHook, "create_model", return_value=None)
+    def test_sagemaker_transform_op_async(
+        self,
+        mock_create_model,
+        mock_list_transform_jobs,
+        mock_transform_job,
+        mock_describe_transform_job,
+        context,
+    ):
         """Assert SageMakerTransformOperatorAsync deferred properly"""
         task = SageMakerTransformOperatorAsync(
             config=CONFIG,
@@ -301,16 +454,165 @@ class TestSagemakerTrainingOperatorAsync:
             (False, SagemakerTrigger, "SagemakerTrigger"),
         ],
     )
+    @mock.patch("astronomer.providers.amazon.aws.operators.sagemaker.SageMakerTrainingOperatorAsync.defer")
+    @mock.patch.object(
+        SageMakerHook,
+        "describe_training_job_with_log",
+        return_value=(
+            ...,
+            {
+                "TrainingJobStatus": "Completed",
+                "ResourceConfig": {"InstanceCount": 1},
+                "TrainingEndTime": datetime(2023, 5, 15),
+                "TrainingStartTime": datetime(2023, 5, 16),
+            },
+            datetime(2023, 5, 16),
+        ),
+    )
+    @mock.patch.object(
+        SageMakerHook,
+        "describe_training_job",
+        return_value={
+            "TrainingJobStatus": "Completed",
+            "ResourceConfig": {"InstanceCount": 1},
+        },
+    )
+    @mock.patch.object(SageMakerHook, "get_conn")
+    @mock.patch.object(SageMakerHook, "create_training_job")
+    @mock.patch.object(sagemaker, "serialize", return_value="")
+    @mock.patch.object(SageMakerHook, "list_training_jobs", return_value=[])
+    def test_sagemaker_training_op_async_complete_before_defer(
+        self,
+        mock_list_training_job,
+        mock_serialize,
+        mock_create_training_job,
+        mock_get_conn,
+        mock_describe_training_job,
+        mock_describe_training_job_with_log,
+        mock_defer,
+        mock_print_log_attr,
+        mock_trigger_class,
+        mock_trigger_name,
+    ):
+        mock_create_training_job.return_value = {
+            "TrainingJobArn": "test_arn",
+            "ResponseMetadata": {"HTTPStatusCode": 200},
+        }
+        task = SageMakerTrainingOperatorAsync(
+            config=TRAINING_CONFIG,
+            task_id=self.TASK_ID,
+            check_if_job_exists=False,
+            print_log=mock_print_log_attr,
+            check_interval=self.CHECK_INTERVAL,
+            max_ingestion_time=self.MAX_INGESTION_TIME,
+        )
+        task.execute(None)
+
+        assert not mock_defer.called
+
+    @pytest.mark.parametrize(
+        "mock_print_log_attr,mock_trigger_class, mock_trigger_name",
+        [
+            (True, SagemakerTrainingWithLogTrigger, "SagemakerTrainingWithLogTrigger"),
+            (False, SagemakerTrigger, "SagemakerTrigger"),
+        ],
+    )
+    @mock.patch("astronomer.providers.amazon.aws.operators.sagemaker.SageMakerTrainingOperatorAsync.defer")
+    @mock.patch.object(
+        SageMakerHook,
+        "describe_training_job_with_log",
+        return_value=(
+            ...,
+            {
+                "TrainingJobStatus": "Failed",
+                "ResourceConfig": {"InstanceCount": 1},
+                "FailureReason": "it just failed",
+            },
+            datetime(2023, 5, 16),
+        ),
+    )
+    @mock.patch.object(
+        SageMakerHook,
+        "describe_training_job",
+        return_value={
+            "TrainingJobStatus": "Failed",
+            "ResourceConfig": {"InstanceCount": 1},
+            "FailureReason": "it just failed",
+        },
+    )
+    @mock.patch.object(SageMakerHook, "get_conn")
+    @mock.patch.object(SageMakerHook, "create_training_job")
+    @mock.patch.object(sagemaker, "serialize", return_value="")
+    @mock.patch.object(SageMakerHook, "list_training_jobs", return_value=[])
+    def test_sagemaker_training_op_async_failed_before_defer(
+        self,
+        mock_list_training_job,
+        mock_serialize,
+        mock_create_training_job,
+        mock_get_conn,
+        mock_describe_training_job,
+        mock_describe_training_job_with_log,
+        mock_defer,
+        mock_print_log_attr,
+        mock_trigger_class,
+        mock_trigger_name,
+    ):
+        mock_create_training_job.return_value = {
+            "TrainingJobArn": "test_arn",
+            "ResponseMetadata": {"HTTPStatusCode": 200},
+        }
+        task = SageMakerTrainingOperatorAsync(
+            config=TRAINING_CONFIG,
+            task_id=self.TASK_ID,
+            check_if_job_exists=False,
+            print_log=mock_print_log_attr,
+            check_interval=self.CHECK_INTERVAL,
+            max_ingestion_time=self.MAX_INGESTION_TIME,
+        )
+        with pytest.raises(AirflowException):
+            task.execute(None)
+
+        assert not mock_defer.called
+
+    @pytest.mark.parametrize(
+        "mock_print_log_attr,mock_trigger_class, mock_trigger_name",
+        [
+            (True, SagemakerTrainingWithLogTrigger, "SagemakerTrainingWithLogTrigger"),
+            (False, SagemakerTrigger, "SagemakerTrigger"),
+        ],
+    )
+    @mock.patch.object(
+        SageMakerHook,
+        "describe_training_job_with_log",
+        return_value=(
+            ...,
+            {
+                "TrainingJobStatus": "InProgress",
+                "ResourceConfig": {"InstanceCount": 1},
+            },
+            datetime(2023, 5, 16),
+        ),
+    )
+    @mock.patch.object(
+        SageMakerHook,
+        "describe_training_job",
+        return_value={
+            "TrainingJobStatus": "InProgress",
+            "ResourceConfig": {"InstanceCount": 1},
+        },
+    )
     @mock.patch.object(SageMakerHook, "get_conn")
     @mock.patch.object(SageMakerHook, "create_training_job")
     @mock.patch.object(sagemaker, "serialize", return_value="")
     @mock.patch.object(SageMakerHook, "list_training_jobs", return_value=[])
     def test_sagemaker_training_op_async(
         self,
-        mock_training_job,
+        mock_list_training_job,
         mock_serialize,
         mock_create_training_job,
-        mock_client,
+        mock_get_conn,
+        mock_describe_training_job,
+        mock_describe_training_job_with_log,
         mock_print_log_attr,
         mock_trigger_class,
         mock_trigger_name,