+++ tests/snowflake/operators/test_snowflake.py
@@ -190,15 +190,111 @@ class TestSnowflakeOperatorAsync:
 
 
 class TestSnowflakeSqlApiOperatorAsync:
-    @pytest.mark.parametrize("mock_sql, statement_count", [(SQL_MULTIPLE_STMTS, 4), (SINGLE_STMT, 1)])
-    @mock.patch(
-        "astronomer.providers.snowflake.hooks.snowflake_sql_api.SnowflakeSqlApiHookAsync.execute_query"
+    @pytest.mark.parametrize(
+        "mock_sql, statement_count, query_ids",
+        [(SQL_MULTIPLE_STMTS, 4, (1, 2, 3, 4)), (SINGLE_STMT, 1, (5,))],
     )
-    def test_snowflake_sql_api_execute_operator_async(self, mock_db_hook, mock_sql, statement_count):
+    @mock.patch("astronomer.providers.snowflake.operators.snowflake.SnowflakeSqlApiOperatorAsync.defer")
+    @mock.patch("requests.session")
+    @mock.patch("astronomer.providers.snowflake.operators.snowflake.SnowflakeSqlApiHookAsync")
+    def test_snowflake_sql_api_execute_operator_async_succeeded_before_defer(
+        self,
+        mock_hook,
+        mock_session,
+        mock_defer,
+        mock_sql,
+        statement_count,
+        query_ids,
+    ):
         """
         Asserts that a task is deferred and an SnowflakeSqlApiTrigger will be fired
         when the SnowflakeSqlApiOperatorAsync is executed.
         """
+        mock_hook.return_value = MagicMock()
+        mock_hook.return_value.query_ids = query_ids
+        mock_hook.return_value.get_request_url_header_params.return_value = ("", "", "")
+
+        mock_session.return_value.__enter__.return_value.get.return_value.__enter__.return_value.status_code = (
+            200
+        )
+
+        operator = SnowflakeSqlApiOperatorAsync(
+            task_id=TASK_ID,
+            snowflake_conn_id=CONN_ID,
+            sql=mock_sql,
+            statement_count=statement_count,
+        )
+
+        operator.execute(create_context(operator))
+
+        assert not mock_defer.called
+
+    @pytest.mark.parametrize(
+        "mock_sql, statement_count, query_ids",
+        [(SQL_MULTIPLE_STMTS, 4, (1, 2, 3, 4)), (SINGLE_STMT, 1, (5,))],
+    )
+    @mock.patch("astronomer.providers.snowflake.operators.snowflake.SnowflakeSqlApiOperatorAsync.defer")
+    @mock.patch("requests.session")
+    @mock.patch("astronomer.providers.snowflake.operators.snowflake.SnowflakeSqlApiHookAsync")
+    def test_snowflake_sql_api_execute_operator_async_failed_before_defer(
+        self,
+        mock_hook,
+        mock_session,
+        mock_defer,
+        mock_sql,
+        statement_count,
+        query_ids,
+    ):
+        """
+        Asserts that a task is deferred and an SnowflakeSqlApiTrigger will be fired
+        when the SnowflakeSqlApiOperatorAsync is executed.
+        """
+        mock_hook.return_value = MagicMock()
+        mock_hook.return_value.query_ids = query_ids
+        mock_hook.return_value.get_request_url_header_params.return_value = ("", "", "")
+
+        mock_session.return_value.__enter__.return_value.get.return_value.__enter__.return_value.status_code = (
+            422
+        )
+
+        operator = SnowflakeSqlApiOperatorAsync(
+            task_id=TASK_ID,
+            snowflake_conn_id=CONN_ID,
+            sql=mock_sql,
+            statement_count=statement_count,
+        )
+
+        with pytest.raises(AirflowException):
+            operator.execute(create_context(operator))
+
+        assert not mock_defer.called
+
+    @pytest.mark.parametrize(
+        "mock_sql, statement_count, query_ids",
+        [(SQL_MULTIPLE_STMTS, 4, (1, 2, 3, 4)), (SINGLE_STMT, 1, (5,))],
+    )
+    @mock.patch("requests.session")
+    @mock.patch("astronomer.providers.snowflake.operators.snowflake.SnowflakeSqlApiHookAsync")
+    def test_snowflake_sql_api_execute_operator_async(
+        self,
+        mock_hook,
+        mock_session,
+        mock_sql,
+        statement_count,
+        query_ids,
+    ):
+        """
+        Asserts that a task is deferred and an SnowflakeSqlApiTrigger will be fired
+        when the SnowflakeSqlApiOperatorAsync is executed.
+        """
+        mock_hook.return_value = MagicMock()
+        mock_hook.return_value.query_ids = query_ids
+        mock_hook.return_value.get_request_url_header_params.return_value = ("", "", "")
+
+        mock_session.return_value.__enter__.return_value.get.return_value.__enter__.return_value.status_code = (
+            202
+        )
+
         operator = SnowflakeSqlApiOperatorAsync(
             task_id=TASK_ID,
             snowflake_conn_id=CONN_ID,